transaction_analysis/pyproject.toml
[project]
name = "transaction-fee-verifier"
version = "0.1.0"
description = "Transaction fee verification agent using Claude"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
]
